+++ internals/logger/logger_test.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 	"os"
 	"testing"
+	"time"
 
 	. "gopkg.in/check.v1"
 	"gopkg.in/tomb.v2"
@@ -108,4 +109,19 @@ func (s *LogSuite) TestMockLoggerReadWriteThreadsafe(c *C) {
 	}
 	err := t.Wait()
 	c.Check(err, IsNil)
+}
+
+func (s *LogSuite) TestAppendTimestamp(c *C) {
+	now := time.Now()
+	c.Assert(string(logger.AppendTimestamp(nil, now)), Equals,
+		now.UTC().Format("2006-01-02T15:04:05.000Z"))
+
+	c.Assert(string(logger.AppendTimestamp(nil, time.Time{})), Equals,
+		"0001-01-01T00:00:00.000Z")
+	c.Assert(string(logger.AppendTimestamp(nil, time.Date(2042, 12, 31, 23, 59, 48, 123_456_789, time.UTC))), Equals,
+		"2042-12-31T23:59:48.123Z")
+	c.Assert(string(logger.AppendTimestamp(nil, time.Date(2025, 8, 9, 1, 2, 3, 4_000_000, time.UTC))), Equals,
+		"2025-08-09T01:02:03.004Z")
+	c.Assert(string(logger.AppendTimestamp(nil, time.Date(2025, 8, 9, 1, 2, 3, 4_999_999, time.UTC))), Equals,
+		"2025-08-09T01:02:03.004Z") // time.Format truncates (not rounds) milliseconds too
 }
